+++ Constants/LocationInfo/WizardTowerSummary.ts
@@ -1,0 +1,27 @@
+import * as Types from "../../Types";
+
+export const WizardTowerSummary: Types.ILocationSummary = {
+  name: `WizardTowerSummary`,
+  displayName: `Wizards' Tower`,
+  Quests: ["Rune Mysteries", "Rune Memories", "Heart of Stone", "Beneath Cursed Tides"],
+  Skills: {
+    Woodcutting: [],
+    Mining: [],
+    Fishing: [],
+    Thieving: {
+      stalls: [],
+      pickpocketing: [],
+    },
+    Farming: [],
+    Firemaking: [],
+    Hunter: [],
+    Divination: [],
+    Archaeology: [],
+    Runecrafting: [],
+    Construction: [],
+    Summoning: [],
+    Agility: [],
+  },
+  Combat: ["lesserdemonwizardtower"],
+  Bosses: [],
+};