+++ src/client/GameComponents/ActivityPanels/MapPanel.tsx
@@ -76,6 +76,7 @@ const MapPanel = (props: Types.MapPanelProps) => {
             {displayTravelButtonJSX(`LumbridgeSwampCave`)}
             {displayTravelButtonJSX(`LumbridgeCatacombs`)}
             {displayTravelButtonJSX(`Draynor`)}
+            {displayTravelButtonJSX(`WizardTower`)}
           </div>
         </div>
       </div>